refactored Secession Manager to adhere to CA
